Mark iana enums as non-exhaustive
If a client of the library happens to `match` on one of these enums,
we don't want the addition of a new value to the range to result in
a build break for the client.  Marking the enums as non-exhaustive
means that clients are required to include a default in their
`match` statements.

https://doc.rust-lang.org/reference/attributes/type_system.html#the-non_exhaustive-attribute