Directly activate alloc feature of ciborium-io
The code relies on the `alloc` feature of the `ciborium-io` crate, but
does not activate it directly. Instead, the feature gets activate by
the dependence on `ciborium`, which does explictly rely on the `alloc`
feature.

The additive nature of features in Cargo means that the `alloc` feature
is therefore always enabled, but it's better to be clear/explicit about
what is needed from dependencies.