header: fix authentication when protected header is zero-length map
COSE allows an empty protected header to be encoded as a zero-length
map, even though the standard encourages encoding empty protected
headers as a zero-length string (RFC 2119 SHOULD according to RFC 9052,
Section 3).

However, according to RFC 9052, Section 4.4, 5.3 and 6.3, even if the
header is encoded as a zero-length map, the structure used for
authentication should not include the empty map if the protected header
is empty ("if there are no protected attributes, a zero-length byte
string is used").

This commit ensures that this behavior is implemented in coset, which
previously did include the zero length map (encoded as h'a0') in
signature calculation.
This previously caused signature verification failures, e.g. when
verifying the CoseSign1 object provided in
https://github.com/cose-wg/Examples/blob/master/sign1-tests/sign-pass-03.json
using coset.