Improve iana_registry macro
For the HeaderBuilder, We had two methods for adding critical label:

    fn add_critical(mut self, param: iana::HeaderParameter) -> Self
    fn add_critical_label(
    	mut self,
	abel: RegisteredLabelWithPrivate<iana::HeaderParameter>) -> Self

The naming of the API itself is confusing, but the reason of having these
two functions is so that we can call add_critical() with
iana::HeaderParamaeter::* directly without wraping it in
RegisteredLabelWithPrivate::Assigned() first.

We have similar requirements for algorithm(), which supports private use
ranges. To make the API clear and easier to use, this commit improves
the iana_registry! macro to provide a from() conversion method. This
allow us to call add_critical() in the following way:

    .add_critical(iana::HeaderParameter::Alg.into())
    .add_critical(RegisteredLabelWithPrivate::PrivateUse(-65537))

Which elimiate the need to have two versions of the add_critical
function. Other methods such as algorithm() which accepts
RegisteredLabelWithPrivate or RegisteredLabel would also automatically
support this usage.