Allow unassigned values for labels
The previous code would reject the use of numeric label values that do
not correspond to an IANA `enum` and which are outside of any defined
private range.

This made it hard to deal with as-yet-unassigned label values from
internet drafts, which are not yet definitive.

So add an `Unassigned(i64)` variant to cover these values.  This also
means that the `UnregisteredIana[NonPrivate]Value` errors can no longer
be emitted.

Fixes #112