Deprecate methods that are easy to misuse (#115)
Various helper methods have documented requirements on their inputs, but
for some of them it's too easy for library users to miss the requirement
and include code that panics on invalid input.

For example, the obvious code to parse and validate a `CoseMac0`:

```
let mac = CoseMac0::from_slice(data);
mac.verify_tag(&aad, |t, d| verify(t, d, k))?;
```

misses the required check that `mac.payload.is_some()`, and so can
panic on invalid input.

So deprecate the helpers that are susceptible to this, in favour of
equivalent methods that explicitly check the prerequisites.